requirements.txt
feedparser>=6.0.11
requests>=2.31.0
websocket-client>=1.8.0
